Make distributed system run from shared folder
It turns out the the distributed akka system (at least the example one
from typesafe) will not run if you are using the default sync between
the virtual machine and the host. I have modified Vagrant file to use
nfs for syncing instead - and it should now work!

Finding this was not easy, so in the future I expect to be titled the
king of debugging!