Handling failure and proof of concept for...
...interacting with the actor system from the outside.

I think that the failed jobs part is ok as is now, though it's not
percistent (which I think it will have to be in the future).

The interactive communitations with the Master is very much of a hack
right now and should be viewed as a proof of concept. Probably there
is some nicer way to do this, but for now I think it's kind of cool. :)