Added packaging capabilities
This includes setting up a simple logger which will write under
/var/log/hercules.log at the moment. Later we might want to hook this up
to logstash or at least make it a rotating log.

I also configured the app main entry to by default try to get the roles
which it should start up (master, watcher, ...) from the
application.conf. That way one or more roles that we want running on the
same machine can be started up together.

Also changed a returns -> return in the documentation (because that is the correct tag to
use).