Improving tracibility
I've added code which will output the Hercules version in the log
everytime actor is starting. In addition to this the "found" file
created when Hercules finds a new runfolder will contain a note on what
Hercules version was used to process it.

All of this info is now pulled from the manifest file, so it will not be
available e.g. when running through sbt, but I decided that this was
good enough, since this data is only critical when going to production
anyways.